Refactor `sf.WSI`; Remove `sf.TMA`
- TMAs have been deprecated. All slides are loaded as WSI.
- In the future, TMA functions (core detection) may be re-implemented as slide-level QC
- Prior base class `SlideLoader` has been folded into `sf.WSI`
