+++ slideflow/slide/utils.py
@@ -1,5 +1,7 @@
 """Utility functions and constants for slide reading."""
 
+import slideflow as sf
+import cv2
 import csv
 import io
 import numpy as np
@@ -7,9 +9,9 @@ import shapely.geometry as sg
 import xml.etree.ElementTree as ET
 
 from PIL import Image, ImageDraw
-from slideflow import errors
+from slideflow import errors, log
 from types import SimpleNamespace
-from typing import Union, List, Tuple, Optional
+from typing import Union, List, Tuple, Optional, Dict
 
 # Constants
 DEFAULT_JPG_MPP = 1
@@ -46,6 +48,162 @@ def OPS_LEVEL_WIDTH(level: int) -> str:
 
 def OPS_LEVEL_DOWNSAMPLE(level: int) -> str:
     return f'openslide.level[{level}].downsample'
+
+# -----------------------------------------------------------------------------
+# Classes
+
+class ROI:
+    """Object container for ROI annotations."""
+
+    def __init__(self, name: str, coordinates: List[Tuple[int, int]] = None) -> None:
+        self.name = name
+        if coordinates is None:
+            self.coordinates = []  # type: List[Tuple[int, int]]
+        else:
+            self.coordinates = coordinates
+
+    def __repr__(self):
+        return f"<ROI (coords={len(self.coordinates)})>"
+
+    def add_coord(self, coord: Tuple[int, int]) -> None:
+        self.coordinates.append(coord)
+
+    def scaled_area(self, scale: float) -> np.ndarray:
+        return np.multiply(self.coordinates, 1/scale)
+
+    def print_coord(self) -> None:
+        for c in self.coordinates:
+            print(c)
+
+    def add_shape(self, shape) -> None:
+        for point in shape:
+            self.add_coord(point)
+
+class QCMask:
+
+    def __init__(self, mask: np.ndarray, filter_threshold: float = 0.6) -> None:
+
+        if not 0 <= filter_threshold <= 1:
+            raise ValueError('filter_threshold must be between 0 and 1')
+        if not isinstance(mask, np.ndarray):
+            raise ValueError('mask must be a numpy array')
+        if not len(mask.shape) == 2:
+            raise ValueError('mask must be a 2D array')
+        if not mask.dtype == bool:
+            raise ValueError('mask must be a boolean array')
+
+        self.mask = mask
+        self.filter_threshold = filter_threshold
+
+    @property
+    def shape(self):
+        return self.mask.shape
+
+# -----------------------------------------------------------------------------
+# Functions
+
+
+def predict(
+    slide: str,
+    model: str,
+    *,
+    stride_div: int = 1,
+    **kwargs
+) -> Tuple[np.ndarray, Optional[np.ndarray]]:
+    """Generate a whole-slide prediction from a saved model.
+
+    Args:
+        slide (str): Path to slide.
+        model (str): Path to saved model trained in Slideflow.
+
+    Keyword args:
+        stride_div (int, optional): Divisor for stride when convoluting
+                across slide. Defaults to 1.
+        roi_dir (str, optional): Directory in which slide ROI is contained.
+            Defaults to None.
+        rois (list, optional): List of paths to slide ROIs. Alternative to
+            providing roi_dir. Defaults to None.
+        roi_method (str): Either 'inside', 'outside', 'auto', or 'ignore'.
+            Determines how ROIs are used to extract tiles.
+            If 'inside' or 'outside', will extract tiles in/out of an ROI,
+            and raise errors.MissingROIError if an ROI is not available.
+            If 'auto', will extract tiles inside an ROI if available,
+            and across the whole-slide if no ROI is found.
+            If 'ignore', will extract tiles across the whole-slide
+            regardless of whether an ROI is available.
+            Defaults to 'auto'.
+        batch_size (int, optional): Batch size for calculating predictions.
+            Defaults to 32.
+        num_threads (int, optional): Number of tile worker threads. Cannot
+            supply both ``num_threads`` (uses thread pool) and
+            ``num_processes`` (uses multiprocessing pool). Defaults to
+            CPU core count.
+        num_processes (int, optional): Number of child processes to spawn
+            for multiprocessing pool. Defaults to None (does not use
+            multiprocessing).
+        enable_downsample (bool, optional): Enable the use of downsampled
+            slide image layers. Defaults to True.
+        img_format (str, optional): Image format (png, jpg) to use when
+            extracting tiles from slide. Must match the image format
+            the model was trained on. If 'auto', will use the format
+            logged in the model params.json. Defaults to 'auto'.
+        generator_kwargs (dict, optional): Keyword arguments passed to
+            the :meth:`slideflow.WSI.build_generator()`.
+        device (torch.device, optional): PyTorch device. Defaults to
+            initializing a new CUDA device.
+
+    Returns:
+        np.ndarray: Predictions for each outcome, with shape = (num_classes, )
+
+        np.ndarray, optional: Uncertainty for each outcome, if the model was
+        trained with uncertainty, with shape = (num_classes,)
+
+    """
+    from slideflow import Heatmap
+    log.info("Calculating whole-slide prediction...")
+    heatmap = Heatmap(slide, model, generate=True, stride_div=stride_div, **kwargs)
+    assert heatmap.predictions is not None
+    preds = heatmap.predictions.reshape(-1, heatmap.predictions.shape[-1])
+    preds = np.ma.masked_where(preds == -99, preds).mean(axis=0).filled()
+    if heatmap.uncertainty is not None:
+        unc = heatmap.uncertainty.reshape(-1, heatmap.uncertainty.shape[-1])
+        unc = np.ma.masked_where(unc == -99, unc).mean(axis=0).filled()
+        return preds, unc
+    else:
+        return preds
+
+
+def log_extraction_params(**kwargs) -> None:
+    """Log tile extraction parameters."""
+
+    if 'whitespace_fraction' not in kwargs:
+        ws_f = DEFAULT_WHITESPACE_FRACTION
+    else:
+        ws_f = kwargs['whitespace_fraction']
+    if 'whitespace_threshold' not in kwargs:
+        ws_t = DEFAULT_WHITESPACE_THRESHOLD
+    else:
+        ws_t = kwargs['whitespace_threshold']
+    if 'grayspace_fraction' not in kwargs:
+        gs_f = DEFAULT_GRAYSPACE_FRACTION
+    else:
+        gs_f = kwargs['grayspace_fraction']
+    if 'grayspace_threshold' not in kwargs:
+        gs_t = DEFAULT_GRAYSPACE_THRESHOLD
+    else:
+        gs_t = kwargs['grayspace_threshold']
+
+    if 'normalizer' in kwargs:
+        log.info(f'Extracting tiles using [magenta]{kwargs["normalizer"]}[/] '
+                 'normalization')
+    if ws_f < 1:
+        log.info('Filtering tiles by whitespace fraction')
+        excl = f'(exclude if >={ws_f*100:.0f}% whitespace)'
+        log.debug(f'Whitespace defined as RGB avg > {ws_t} {excl}')
+    if gs_f < 1:
+        log.info('Filtering tiles by grayspace fraction')
+        excl = f'(exclude if >={gs_f*100:.0f}% grayspace)'
+        log.debug(f'Grayspace defined as HSV avg < {gs_t} {excl}')
 
 
 def draw_roi(
@@ -233,7 +391,7 @@ def _find_translation_matrix(
     # Use findTransformECC to compute the transformation
     _, warp_matrix = cv2.findTransformECC(im2_gray, im1_gray, warp_matrix, warp_mode, criteria)
 
-    return warp_matrix
+    return warp_matrix  # type: ignore
 
 
 def align_image(im1: np.ndarray, im2: np.ndarray) -> np.ndarray:
@@ -286,6 +444,7 @@ def align_by_translation(
         return alignment, mse
     else:
         return alignment
+
 
 def compute_alignment_mse(
     imageA: np.ndarray,
@@ -363,3 +522,71 @@ def z_on_plane(x, y, centroid, normal):
 
     z = cz + (nx * (cx - x) + ny * (cy - y)) / nz
     return z
+
+
+def calc_alignment(c, us, them, n=None):
+    idx, (x, y, xi, yi) = c
+    our_tile = us[xi, yi]
+    try:
+        their_tile = them[xi, yi]
+    except IndexError:
+        return None, c
+    if our_tile is None or their_tile is None:
+        return None, c
+    if n is not None:
+        our_tile = n.transform(our_tile[:, :, 0:3])
+        their_tile = n.transform(their_tile[:, :, 0:3])
+    try:
+        rough_alignment = sf.slide.utils._find_translation_matrix(their_tile, our_tile, h=50, search_window=53)
+    except cv2.error:
+        rough_alignment = None
+        log.debug("Initial rough alignment failed at x={}, y={} (grid {}, {})".format(
+            x, y, xi, yi
+        ))
+    else:
+        log.debug("Initial rough alignment complete at x={}, y={} (grid {}, {}): {}".format(
+            x, y, xi, yi, (int(np.round(-rough_alignment[0, 2])), int(np.round(-rough_alignment[1, 2])))
+        ))
+    try:
+        return align_by_translation(their_tile, our_tile, round=True, warp_matrix=rough_alignment), c
+    except errors.AlignmentError as e:
+        return 'error', c
+
+# -----------------------------------------------------------------------------
+# Internals
+
+def _update_kw_with_defaults(kwargs) -> Dict:
+    """Updates a set of keyword arguments with default extraction values.
+    for whitepsace/grayspace filtering.
+    """
+    if kwargs['whitespace_fraction'] is None:
+        kwargs['whitespace_fraction'] = DEFAULT_WHITESPACE_FRACTION
+    if kwargs['whitespace_threshold'] is None:
+        kwargs['whitespace_threshold'] = DEFAULT_WHITESPACE_THRESHOLD
+    if kwargs['grayspace_fraction'] is None:
+        kwargs['grayspace_fraction'] = DEFAULT_GRAYSPACE_FRACTION
+    if kwargs['grayspace_threshold'] is None:
+        kwargs['grayspace_threshold'] = DEFAULT_GRAYSPACE_THRESHOLD
+    if kwargs['img_format'] is None:
+        kwargs['img_format'] = 'jpg'
+    return kwargs
+
+
+def _polyArea(x: List[float], y: List[float]) -> float:
+    return 0.5*np.abs(np.dot(x, np.roll(y, 1))-np.dot(y, np.roll(x, 1)))
+
+
+def _convert_img_to_format(image: np.ndarray, img_format: str) -> str:
+    if img_format.lower() == 'png':
+        return cv2.imencode(
+            '.png',
+            cv2.cvtColor(image, cv2.COLOR_RGB2BGR)
+        )[1].tobytes()
+    elif img_format.lower() in ('jpg', 'jpeg'):
+        return cv2.imencode(
+            '.jpg',
+            cv2.cvtColor(image, cv2.COLOR_RGB2BGR),
+            [int(cv2.IMWRITE_JPEG_QUALITY), 100]
+        )[1].tostring()
+    else:
+        raise ValueError(f"Unknown image format {img_format}")
